Force-adding a "Penn State Extension Internal Store" category if the internal store is selected.
This is implemented on the @indexer method, and my intent was to put that data only in the catalog.  However, that appears to also be reflected on the product
itself, so this is a bonus.

Created two utility methods that we can re-use to get the id of the internal store, and see if the product is in the internal store.