Added an @expensive decorator to denote that an adapter method is 'expensive' in terms of calcluations required, and we can skip that if we don't need it.  This is True by default, except for the PloneSite API view.
Enabled that for the 'related products' field.

Reason for this change is that pulling many products having the 'related_products' field takes lots of extra time, since it does a bunch of
calculations.  If we don't show this field for the main @@api view, a big feed will take less time to generate.